+++ lib/int/int.dart
@@ -107,4 +107,14 @@ extension Int_ on int {
   Duration get days {
     return Duration(days: this);
   }
+
+  /// Returns a [IntTween] from [this] value to the given [end] value
+  ///
+  /// Example:
+  /// ```dart
+  /// 100.tweenTo(200) // Tween(begin: 100, end: 200);
+  /// ```
+  IntTween tweenTo(int end) {
+    return IntTween(begin: this, end: end);
+  }
 }